Refactor: Extract shared instance preparation logic
Extracted common instance validation and startup logic from ShellCommand
and RunCommand into a new InstancePrep utility class.

Changes:
- New InstancePrep.prepareInstance() method handles:
  - Existence check
  - Parent template validation
  - Network mode and proxy health checks
  - Instance startup if stopped
  - GUI passthrough health check
  - CA certificate mismatch fixes

- ShellCommand: Reduced from 75 to 38 lines by using InstancePrep
- RunCommand: Reduced from 175 to 123 lines by using InstancePrep
- Eliminates ~80 lines of duplicated code

The TUI (ListCommand) is intentionally not modified as it already had
duplicated logic and has different requirements for error handling
(status messages vs exit codes) due to its persistent interactive loop.

All 655 tests pass.
