+++ src/main/java/dev/incusspawn/command/RunCommand.java
@@ -2,14 +2,6 @@ package dev.incusspawn.command;
 
 import dev.incusspawn.RuntimeServices;
 import dev.incusspawn.config.ImageDef;
-import dev.incusspawn.config.NetworkMode;
-import dev.incusspawn.config.HostResourceSetup;
-import dev.incusspawn.incus.BridgeSubnetCheck;
-import dev.incusspawn.incus.FirewalldCheck;
-import dev.incusspawn.incus.Metadata;
-import dev.incusspawn.lifecycle.GuiPassthrough;
-import dev.incusspawn.proxy.CertificateAuthority;
-import dev.incusspawn.proxy.ProxyHealthCheck;
 import dev.incusspawn.tool.ActionResolver;
 import dev.incusspawn.tool.ToolAction;
 import dev.incusspawn.tool.YamlToolAction;
@@ -37,37 +29,10 @@ public class RunCommand extends BaseCommand {
     protected CommandResult doExecute() throws Exception {
         var incus = RuntimeServices.incus();
 
-        if (!incus.exists(name)) {
-            System.err.println("Error: no instance named '" + name + "' found.");
-            System.err.println("Run 'isx list' to see available instances.");
+        var parent = InstancePrep.prepareInstance(incus, name);
+        if (parent == null) {
             return CommandResult.valueOf(1);
         }
-
-        // Validate parent template before any side effects
-        var parent = incus.configGet(name, Metadata.PARENT);
-        if (parent == null || parent.isEmpty()) {
-            System.err.println("Error: instance '" + name + "' has no parent template.");
-            System.err.println("This does not appear to be an incus-spawn managed instance.");
-            return CommandResult.valueOf(1);
-        }
-
-        var networkMode = incus.configGet(name, Metadata.NETWORK_MODE);
-        if (!NetworkMode.AIRGAP.name().equals(networkMode)) {
-            if (!ProxyHealthCheck.checkOrWarn(incus)) return CommandResult.valueOf(1);
-            BridgeSubnetCheck.warnIfConflict(incus);
-            FirewalldCheck.warnIfNotRunning();
-            fixCaMismatch(incus, name);
-        }
-
-        // Start if stopped
-        if ("Stopped".equalsIgnoreCase(incus.getInstanceStatus(name))) {
-            System.out.println("Starting " + name + "...");
-            HostResourceSetup.removeStaleDevices(incus, name);
-            incus.start(name);
-            incus.waitForReady(name);
-        }
-
-        GuiPassthrough.checkGuiHealth(incus, name);
 
         // Build the action resolver
         var toolDefLoader = RuntimeServices.toolDefLoader();
@@ -155,21 +120,5 @@ public class RunCommand extends BaseCommand {
         }
 
         return CommandResult.SUCCESS;
-    }
-
-    private void fixCaMismatch(dev.incusspawn.incus.IncusClient incus, String container) {
-        // Ensure the container is running so we can push the cert
-        if ("Stopped".equalsIgnoreCase(incus.getInstanceStatus(container))) {
-            HostResourceSetup.removeStaleDevices(incus, container);
-            incus.start(container);
-            incus.waitForReady(container);
-        }
-
-        if (CertificateAuthority.fixContainerCaIfNeeded(incus, container)) {
-            var sep = "\033[33m" + "─".repeat(60) + "\033[0m";
-            System.err.println(sep);
-            System.err.println("\033[1;33mCA certificate mismatch\033[0m — updated automatically.");
-            System.err.println(sep);
-        }
     }
 }
